Throw it all away and start again!
I learned a lot from dabbling in a C# version of this library, so I'm going
to try and apply what I learned to this library. Unfortunately that means
throwing a lot out, and as of this commit most of it is not put back.

I will not merge this commit to master by itself, so this commit message
should not be seen by anyone except myself and the really curious. :)