Add untested live data fetching ported over from pre-rewrite.
I haven't tested this on an actual ECU yet, but it's mostly a copy-paste
from before I started the rewrite. The only real changes are outside of the
function (where we change the behaviour of when to wait for a prompt to make
it more resilient), and we also remove the spaces so all the regexes had to
be altered in `elm327/pids.py`... it's possible I missed one though.